payments/src/events/listeners/order_created_listener.ts
import {Listener, OrderCreatedEvent, Subjects} from '@ac-tickets/common'
import {queueGroupName} from './queue_group_name'
import { Message } from 'node-nats-streaming'
import { Order } from '../../models/order'

export class OrderCreatedListener extends Listener<OrderCreatedEvent>{
  readonly subject = Subjects.OrderCreated
  readonly queueGroupName = queueGroupName

  async onMessage(data: OrderCreatedEvent['data'], msg: Message) {
    const order = Order.build({
      id: data.id,
      price: data.ticket.price,
      status: data.status,
      userId: data.userId,
      version: data.version,
    })
    await order.save()
    
    msg.ack()
  }
}